+++ exposan/htl/system_no_P.py
@@ -1774,36 +1774,3 @@ def get_LCA_diesel():
 @metric(name='GWP_sludge',units='kg CO2/ton dry sludge',element='LCA')
 def get_LCA_sludge():
     return lca_sludge.get_total_impacts()['GlobalWarming']/raw_wastewater.F_vol/_m3perh_to_MGD/WWTP.ww_2_dry_sludge/(sys_no_P.operating_hours/24)/lca_sludge.lifetime
-
-#%%
-np.random.seed(3221)
-samples = model.sample(N=1000, rule='L')
-model.load_samples(samples)
-model.evaluate()
-model.table
-
-#%%
-def organize_results(model, path):
-    idx = len(model.parameters)
-    parameters = model.table.iloc[:, :idx]
-    results = model.table.iloc[:, idx:]
-    percentiles = results.quantile([0, 0.05, 0.25, 0.5, 0.75, 0.95, 1])
-    with pd.ExcelWriter(path) as writer:
-        parameters.to_excel(writer, sheet_name='Parameters')
-        results.to_excel(writer, sheet_name='Results')
-        percentiles.to_excel(writer, sheet_name='Percentiles')
-organize_results(model, 'example_model.xlsx')
-
-#%%
-fig, ax = qs.stats.plot_uncertainties(model)
-fig
-
-#%%
-fig, ax = qs.stats.plot_uncertainties(model, x_axis=model.metrics[0], y_axis=model.metrics[1],
-                                      kind='kde-kde', center_kws={'fill': True})
-fig
-
-#%%
-r_df, p_df = qs.stats.get_correlations(model, kind='Spearman', nan_policy='omit')
-fig, ax = qs.stats.plot_correlations(r_df)
-fig
